Move manager to seperate module
